requirements.txt
parallel-ssh==2.12.0
watchdog==3.0.0
PyYAML==6.0.1